requirements.txt
psycopg[binary]>=3.2.0
python-dotenv==1.0.0

